fix: dial relay when we are dialed via it but have no reservation
Fixes an edge case where we receive an incoming connection attempt
from a peer via a relay that we did not know we had a reservation
on.

This could happen if we restart or otherwise lose context.

When this happens, attempt to listen on a p2p-circuit version of
the incoming connection remote addr - this will cause us to refresh
our reservation on the remote relay.

Fixes recent interop test instability where the test configures a
relay reservation on behalf of a node which then has no knowledge of
it's reservations so tells identify peers that it has no listen
addresses.
